+++ apps/block_scout_web/assets/js/lib/smart_contract/functions.js
@@ -1,6 +1,7 @@
 import $ from 'jquery'
 import { getContractABI, getMethodInputs, prepareMethodArgs } from './common_helpers'
-import { walletEnabled, connectToWallet, shouldHideConnectButton, callMethod, queryMethod } from './write'
+import { queryMethod, callMethod } from './interact'
+import { walletEnabled, connectToWallet, disconnectWallet, web3ModalInit } from './connect.js'
 import '../../pages/address'
 import * as Sentry from '@sentry/browser'
 
@@ -17,37 +18,9 @@ const loadFunctions = (element) => {
     response => $element.html(response)
   )
     .done(function () {
-      const $connect = $('[connect-metamask]')
-      const $connectTo = $('[connect-to]')
-      const $connectedTo = $('[connected-to]')
-      const $reconnect = $('[re-connect-metamask]')
-
-      window.ethereum && window.ethereum.on('accountsChanged', function (accounts) {
-        if (accounts.length === 0) {
-          showConnectElements($connect, $connectTo, $connectedTo)
-        } else {
-          showConnectedToElements($connect, $connectTo, $connectedTo, accounts[0])
-        }
-      })
-
-      shouldHideConnectButton()
-        .then(({ shouldHide, account }) => {
-          if (shouldHide && account) {
-            showConnectedToElements($connect, $connectTo, $connectedTo, account)
-          } else if (shouldHide) {
-            hideConnectButton($connect, $connectTo, $connectedTo)
-          } else {
-            showConnectElements($connect, $connectTo, $connectedTo)
-          }
-        })
-
-      $connect.on('click', () => {
-        connectToWallet()
-      })
-
-      $reconnect.on('click', () => {
-        connectToWallet()
-      })
+      document.querySelector('[connect-wallet]') && document.querySelector('[connect-wallet]').addEventListener('click', connectToWallet)
+      document.querySelector('[disconnect-wallet]') && document.querySelector('[disconnect-wallet]').addEventListener('click', disconnectWallet)
+      web3ModalInit(connectToWallet)
 
       $('[data-function]').each((_, element) => {
         readWriteFunction(element)
@@ -74,30 +47,6 @@ const loadFunctions = (element) => {
     .fail(function (response) {
       $element.html(response.statusText)
     })
-}
-
-function showConnectedToElements ($connect, $connectTo, $connectedTo, account) {
-  $connectTo.addClass('hidden')
-  $connect.removeClass('hidden')
-  $connectedTo.removeClass('hidden')
-  setConnectToAddress(account)
-}
-
-function setConnectToAddress (account) {
-  const $connectedToAddress = $('[connected-to-address]')
-  $connectedToAddress.html(`<a href='/address/${account}'>${account}</a>`)
-}
-
-function showConnectElements ($connect, $connectTo, $connectedTo) {
-  $connectTo.removeClass('hidden')
-  $connect.removeClass('hidden')
-  $connectedTo.addClass('hidden')
-}
-
-function hideConnectButton ($connect, $connectTo, $connectedTo) {
-  $connectTo.removeClass('hidden')
-  $connect.addClass('hidden')
-  $connectedTo.addClass('hidden')
 }
 
 const readWriteFunction = (element) => {