+++ apps/block_scout_web/lib/block_scout_web/csp_header.ex
@@ -11,7 +11,7 @@ defmodule BlockScoutWeb.CSPHeader do
   def call(conn, _opts) do
     Controller.put_secure_browser_headers(conn, %{
       "content-security-policy" => "\
-        connect-src 'self' #{websocket_endpoints(conn)} *.sentry.io *.google-analytics.com https://stats.g.doubleclick.net/ *.poa.network/ https://request-global.czilladx.com/ *.googlesyndication.com/ https://raw.githubusercontent.com/trustwallet/assets/; \
+        connect-src 'self' #{websocket_endpoints(conn)} wss://*.bridge.walletconnect.org/ *.sentry.io *.google-analytics.com https://stats.g.doubleclick.net/ *.poa.network/ https://request-global.czilladx.com/ *.googlesyndication.com/ https://raw.githubusercontent.com/trustwallet/assets/ https://registry.walletconnect.org/data/wallets.json;\
         default-src 'self';\
         script-src 'self' 'unsafe-inline' 'unsafe-eval' *.googletagmanager.com *.google-analytics.com https://www.google.com https://www.gstatic.com https://hcaptcha.com https://assets.hcaptcha.com https://newassets.hcaptcha.com https://coinzillatag.com *.googlesyndication.com https://adservice.google.com https://adservice.google.ru *.googletagservices.com *.googleadservices.com; \
         style-src 'self' 'unsafe-inline' 'unsafe-eval' https://fonts.googleapis.com;\